Reword the error that is logged when skipping choice types on slice extensions.
I believe we are skipping it in the right place, but the error message (and the fact that we were previously raising an error) was misleading. Updated the error message now.

Also adds more test coverage in `fhir_path_test.py` for regex validation. This was already tested in the minion tests.

PiperOrigin-RevId: 511861582