Re-validate the operands and re-populate node's return type once replacing the operand
For most of the nodes, the return_type is populated from their operands. Make sure operands are re-validated and then re-populate the return type after replacing the operands.

This change is for preparing the forEach feature in FHIRViews, where the operand node will be replaced by itself with SCALAR return_type.

PiperOrigin-RevId: 572585233